added minimum static style to form components added simple style to the indicator form refractored code to deprecate simpleForm component
