+++ app/src/Components/FormInput/CheckboxInput.js
@@ -16,7 +16,7 @@ const CheckboxInput = ({
   ...props
 }) => {
   return (
-    <FormControlLabel
+    <FormControlLabel style={{width: '250px', margin: '15px'}}
       control={
         <React.Fragment>
           <Checkbox
