+++ app/src/Components/FormInput/DateInput.js
@@ -17,7 +17,7 @@ const DateInput = ({
   ...props
 }) => {
   return (
-    <FormControlLabel
+    <FormControlLabel style={{width: '250px', margin: '15px'}}
       control={
         <React.Fragment>
           <TextField
