+++ app/src/Components/FormInput/TextInput.js
@@ -19,7 +19,7 @@ const TextInput = ({
   ...props
 }) => {
   return (
-    <FormControlLabel
+    <FormControlLabel style={{width: '250px', margin: '15px'}}
       control={
         <React.Fragment>
           <TextField
